drivers/firmware/ukacpi: Add MADT-based CPU idx/id map filler
Define an early boot routine based on the moved implementation of
`uk_plat_native_lcpu_mp_init` that just parses the MADT filling in
per-CPU ID/IDX. For correctness, the `uk_acpi_init` should also
be registered in the boot early table, with a preceding priority.

To correctly renounce ownership of this responsibility from the
platform's lcpu pm driver, we must also register an early handler
for the SIPI vector allocation.

Finally, we no longer need this "lcpu_mp_init" routine in the PAL
or have it exported by the platform driver, so remove it along with
any related registrations.

Checkpatch-Ignore: AVOID_EXTERNS
