+++ drivers/firmware/ukacpi/Makefile.uk
@@ -7,3 +7,7 @@ CXXINCLUDES-$(CONFIG_LIBUKACPI) += -I$(LIBUKACPI_BASE)/include
 LIBUKACPI_CINCLUDES-y += -I$(UK_PLAT_COMMON_BASE)/include
 
 LIBUKACPI_SRCS-$(CONFIG_LIBUKACPI) += $(LIBUKACPI_BASE)/acpi.c
+
+ifeq ($(CONFIG_ARCH_X86_64), y)
+LIBUKACPI_SRCS-$(CONFIG_LIBUKACPI_MADT_FILL_CPU_IDMAP) += $(LIBUKACPI_BASE)/arch/x86_64/madt.c
+endif
