+++ drivers/firmware/ukacpi/include/uk/acpi/prio.h
@@ -1,0 +1,14 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+/* Copyright (c) 2026, Unikraft GmbH and The Unikraft Authors.
+ * Licensed under the BSD-3-Clause License (the "License").
+ * You may not use this file except in compliance with the License.
+ */
+
+#ifndef __UK_ACPI_PRIO_H__
+#define __UK_ACPI_PRIO_H__
+
+#include <uk/prio.h>
+
+#define UK_ACPI_INIT_PRIO				UK_PRIO_EARLIEST
+
+#endif /* __UK_ACPI_PRIO_H__ */
